解决OpengGl 0x502不断打印的bug，主要是float改为int，如GLES20.glUniform1i(textureHandle,0)

+++ media/src/main/java/com/sky/media/kit/render/FaceBuffingRender.kt
@@ -150,6 +150,6 @@ class FaceBuffingRender : BaseRender(), IAdjustable {
     }
 
     override fun adjust(i: Int, i2: Int, i3: Int) {
-        mBuffingLevel = (i - i2).toFloat() * 1.0f / (i3 - i2).toFloat()
+        mBuffingLevel = (i - i2) * 1.0f / (i3 - i2)
     }
 }
